drizzle_unpack_string() integer-width safety. The size of a buffer in our memory space is a 'size_t'. Fail if the string extends past the limits of the buffer we're reading it from (should only happen if there's a protocol error).

+++ libdrizzle/pack.h
@@ -85,9 +85,10 @@ void drizzle_unpack_datetime(drizzle_field_t field, size_t length, unsigned char
 
 /**
  * Unpack length-encoded string.
+ * max_size is a size_t because it describes the size of the in-core 'buffer'.
  */
 drizzle_return_t drizzle_unpack_string(drizzle_st *con, char *buffer,
-                                       uint64_t max_size);
+                                       size_t max_size);
 
 /**
  * Pack user, scramble, and db.
